requirements.txt
monai==1.4.0
numpy==2.2.5
pandas==2.2.3
PyYAML==6.0.2
PyYAML==6.0.2
torch==2.7.0+cu128
